requirements.txt
Django==2.2.3
lxml==4.3.4
pkg-resources==0.0.0
python-docx==0.8.10
pytz==2019.1
sqlparse==0.3.0
